PythonServer/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
groq==0.4.2
SpeechRecognition==3.10.1
pydub==0.25.1
pyaudio==0.2.14
gTTS==2.5.1
python-dotenv==1.0.1
aiofiles==23.2.1
